+++ src/App.tsx
@@ -36,7 +36,7 @@ import CreatePurchaseRequest from './screens/Inventory/CreatePurchaseRequest';
 import InventoryMain from './screens/Inventory/InventoryMain';
 import InventoryProfile from './screens/Inventory/InventoryProfile';
 import PurchaseRequest from './screens/Inventory/PurchaseRequest';
-import PurchaseRequests from './screens/Inventory/PurchaseRequests';
+import PurchaseRequestsMain from './screens/Inventory/PurchaseRequestsMain';
 import Login from './screens/Login';
 import Maintenance from './screens/Maintenance';
 import { theme } from './styles/ThemeStyles';
@@ -74,7 +74,7 @@ function App() {
       <AuthenticatedRoute path="/inventory" component={InventoryMain} exact/>
       <AuthenticatedRoute path="/inventory/create" component={AddInventory} />
       <AuthenticatedRoute path="/inventory/item" component={InventoryProfile} /> 
-      <AuthenticatedRoute path="/inventory/purchase-requests" component={PurchaseRequests} exact/> 
+      <AuthenticatedRoute path="/inventory/purchase-requests" component={PurchaseRequestsMain} exact/> 
       <AuthenticatedRoute path="/inventory/purchase-requests/create" component={CreatePurchaseRequest} /> 
       <AuthenticatedRoute path="/inventory/purchase-requests/purchase-request" component={PurchaseRequest} /> 
       <AuthenticatedRoute path="/inventory/updates/create" component={CreateInventoryUpdate} />
